requirements.txt
astroid==2.4.2
colorama==0.4.3
Django==2.2
django-cleanup==5.1.0
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-filter==2.4.0
django-model-utils==4.0.0
django-notifications-hq==1.6.0
djangorestframework==3.12.1
gunicorn==20.0.4
isort==5.5.3
jsonfield==3.1.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.0.0
psycopg2==2.8.6
pylint==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
swapper==1.1.2.post1
toml==0.10.1
wrapt==1.12.1
